extract grammar linting
